requirements.txt
Django>=5.2.7
djangorestframework>=3.15.0
django-cors-headers>=4.5.0
